Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/content-security-policy/inheritance/inherited-csp-list-modifications-are-local-expected.txt

CONSOLE MESSAGE: Refused to load http://localhost:8800/content-security-policy/support/fail.png because it does not appear in the img-src directive of the Content Security Policy.


Harness Error (TIMEOUT), message = null

PASS Test that parent document image loads
PASS Test that embedded iframe document image does not load
NOTRUN Test that spv event is fired
